% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aleutians}
\alias{aleutians}
\title{Bathymetric data for the Aleutians (Alaska)}
\usage{
data(aleutians)
}
\value{
A text file.
}
\description{
Bathymetric matrix of class \code{bathy} created from NOAA GEODAS data.
}
\details{
Data imported from the NOAA Grid Extract webpage (\url{https://www.ncei.noaa.gov/maps/grid-extract/}) and transformed into an object of class \code{bathy} by \code{mar_as_bathy}.
}
\examples{
# load celt data
data(aleutians)

# class "bathy"
class(aleutians)
summary(aleutians)

# test mar_plot_bathy
plot(aleutians,image = TRUE,
     bpal = list(c(0,max(aleutians),"grey"),
                 c(min(aleutians),0,"darkblue","lightblue")),
     land = TRUE, lwd = 0.1, axes = FALSE)
mar_antimeridian_box(aleutians, 10)
}
\seealso{
\code{\link{mar_as_bathy}}, \code{\link{mar_read_bathy}}, \code{\link{mar_antimeridian_box}}
}
\author{
see \url{https://www.ncei.noaa.gov/maps/grid-extract/}
}
