% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_bathy.R
\name{as_bathy}
\alias{as_bathy}
\title{Convert to bathymetric data in an object of class bathy}
\usage{
as_bathy(x)
}
\arguments{
\item{x}{Three-column data frame with longitude, latitude and depth values.}
}
\value{
The output of \code{as_bathy} is a matrix of class \code{bathy}, with
longitude stored in row names and latitude stored in column names.
}
\description{
Converts a three-column data frame containing longitude, latitude and depth
values to a matrix of class \code{bathy}.
}
\details{
The first column is interpreted as longitude, the second as latitude, and
the third as depth or elevation.
}
\examples{
xyz <- data.frame(
  lon = rep(c(-5, -4, -3), each = 3),
  lat = rep(c(48, 49, 50), times = 3),
  depth = c(-80, -70, -60, -120, -110, -100, -160, -150, -140)
)

bathy <- as_bathy(xyz)
class(bathy)
summarise_bathy(bathy)
}
\seealso{
\code{\link{summarise_bathy}}, \code{\link{read_bathy}},
\code{\link{as_xyz}}, \code{\link{bathy_to_tbl}}, \code{\link{tbl_to_bathy}}.
}
\author{
Benoit Simon-Bouhet
}
