% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celt}
\alias{celt}
\title{Bathymetric data for the North Est Atlantic}
\usage{
data(celt)
}
\value{
A text file.
}
\description{
Bathymetric matrix of class \code{bathy} created from NOAA GEODAS data.
}
\details{
Data imported from the NOAA Grid Extract webpage (\url{https://www.ncei.noaa.gov/maps/grid-extract/}) and transformed into an object of class \code{bathy} by \code{mar_as_bathy}.
}
\examples{
# load celt data
data(celt)

# class "bathy"
class(celt)
summary(celt)

# test mar_plot_bathy
plot(celt, deep=-300, shallow=-50, step=25)
}
\seealso{
\code{\link{mar_as_bathy}}, \code{\link{mar_read_bathy}}
}
\author{
see \url{https://www.ncei.noaa.gov/maps/grid-extract/}
}
