% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bathy.R
\name{check_bathy}
\alias{check_bathy}
\title{Sort bathymetric data matrix by increasing latitude and longitude}
\usage{
check_bathy(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
The output of \code{check_bathy} is an ordered matrix.
}
\description{
Reads a bathymetric data matrix and orders its rows and columns by increasing latitude and longitude.
}
\details{
\code{check_bathy} allows to sort rows and columns by increasing latitude and longitude, which is necessary for ploting with the function \code{image} (package \code{graphics}). \code{check_bathy} is used within the \code{marmap} functions \code{read_bathy} and \code{as_bathy} (it is also used in \code{get_noaa} through \code{as_bathy}).
}
\examples{
matrix(1:100, ncol=5, dimnames=list(20:1, c(3,2,4,1,5))) -> a
check_bathy(a)
}
\seealso{
\code{\link{read_bathy}}, \code{\link{as_bathy}}, \code{\link{get_noaa}}
}
\author{
Eric Pante
}
