% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_bathy.R
\name{collate_bathy}
\alias{collate_bathy}
\title{Collates two bathy matrices with data from either sides of the antimeridian}
\usage{
collate_bathy(east,west)
}
\arguments{
\item{east}{matrix of class \code{bathy} with eastern data (West of antimeridian)}

\item{west}{matrix of class \code{bathy} with western data (East of antimeridian)}
}
\value{
A single matrix of class \code{bathy}. When plotting collated data with
\code{\link{geom_bathy}}, use \code{antimeridian = TRUE} to display western
longitude labels beyond 180 degrees.
}
\description{
Collates two bathy matrices, one with longitude 0 to 180 degrees East, and the other with longitude 0 to 180 degrees West
}
\details{
This function is used internally by import functions when data are
downloaded from both sides of the antimeridian line (180 degrees longitude).
If, for example, data are downloaded for longitudes 170E-180 and 180-170W,
\code{collate_bathy()} creates a single matrix of class \code{bathy} with a
coordinate system going from 170 to 190 degrees longitude.

\code{get_noaa()} deals with data from both sides of the antimeridian and does not need further processing with \code{collate_bathy()}.
}
\examples{
east <- as_bathy(data.frame(
  lon = rep(c(178, 179), each = 2),
  lat = rep(c(10, 11), times = 2),
  depth = c(-10, -20, -30, -40)
))
west <- as_bathy(data.frame(
  lon = rep(c(-180, -179), each = 2),
  lat = rep(c(10, 11), times = 2),
  depth = c(-50, -60, -70, -80)
))

collate_bathy(east, west)
}
\seealso{
\code{\link{get_noaa}}, \code{\link{get_gebco}}, \code{\link{geom_bathy}},
\code{\link{summary_bathy}}
}
\author{
Eric Pante
}
