% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{florida}
\alias{florida}
\title{Bathymetric data around Florida, USA}
\usage{
data(florida)
}
\value{
A bathymetric object of class \code{bathy} with 539 rows and 659 columns.
}
\description{
Bathymetric object of class \code{bathy} created from NOAA GEODAS data.
}
\details{
Data imported from the NOAA Grid Extract webpage (\url{https://www.ncei.noaa.gov/maps/grid-extract/}) and transformed into an object of class \code{bathy} by \code{mar_read_bathy}.
}
\examples{
# load florida data
data(florida)

# class "bathy"
class(florida)
summary(florida)

# test mar_plot_bathy
plot(florida,asp=1)
plot(florida,asp=1,image=TRUE,drawlabels=TRUE,land=TRUE,n=40)
}
\seealso{
\code{\link{mar_plot_bathy}}, \code{\link{mar_summary_bathy}}
}
\author{
see \url{https://www.ncei.noaa.gov/maps/grid-extract/}
}
