% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sf_antimeridian.R
\name{geom_sf_antimeridian}
\alias{geom_sf_antimeridian}
\title{Plot sf bathymetry around the antimeridian}
\usage{
geom_sf_antimeridian(
  mapping = NULL,
  data = NULL,
  stat = "sf",
  position = "identity",
  ...,
  x_breaks = ggplot2::waiver(),
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes}}.}

\item{data}{Data to display in this layer. Defaults to the data inherited
from \code{\link[ggplot2:ggplot]{ggplot2::ggplot}}.}

\item{stat}{Statistical transformation used by
\code{\link[ggplot2:geom_sf]{ggplot2::geom_sf}}. Defaults to \code{"sf"}.}

\item{position}{Position adjustment. Defaults to \code{"identity"}.}

\item{...}{Additional arguments passed to
\code{\link[ggplot2:geom_sf]{ggplot2::geom_sf}}.}

\item{x_breaks}{Breaks used for the x axis. The default lets ggplot2 choose
breaks automatically. This only affects the longitude axis labels added
by \code{\link{coord_sf_antimeridian}}.}

\item{na.rm, show.legend, inherit.aes}{Arguments passed to
\code{\link[ggplot2:geom_sf]{ggplot2::geom_sf}}.}
}
\value{
A list of ggplot2 components that can be added to a plot with \code{+}.
}
\description{
Adds an \code{sf} layer with longitude axis labels adapted for data stored
in the 0-360 longitude convention around the antimeridian.
}
\examples{
\dontrun{
anti |>
  as_sf() |>
  ggplot2::ggplot() +
  geom_sf_antimeridian(ggplot2::aes(color = depth))
}
}
\seealso{
\code{\link{as_sf}},
\code{\link[ggplot2:geom_sf]{ggplot2::geom_sf}},
\code{\link{coord_sf_antimeridian}}
}
