% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample.R
\name{get_sample}
\alias{get_sample}
\title{Get sample data by clicking on a map}
\usage{
get_sample(mat, sample, col.lon, col.lat, \dots)
}
\arguments{
\item{mat}{bathymetric data matrix of class \code{bathy}, imported using \code{read_bathy} (no default)}

\item{sample}{data.frame containing sampling information (at least longitude and latitude) (no default)}

\item{col.lon}{column number of data frame \code{sample} containing longitude information (no default)}

\item{col.lat}{column number of data frame \code{sample} containing latitude information (no default)}

\item{\dots}{further arguments to be passed to \code{\link{rect}} for drawing a box around the selected area}
}
\value{
a dataframe of the elements of \code{sample} present within the area selected
}
\description{
Outputs sample information based on points selected by clicking on a map
}
\details{
\code{get_sample} allows the user to get sample data by clicking on a map created with \code{plot_bathy}. This function uses the \code{locator} function (\code{graphics} package). After creating a map with \code{plot_bathy}, the user can  click twice on the map to delimit an area (for example, lower left and upper right corners of a rectangular area of interest), and get a dataframe corresponding to the \code{sample} points present within the selected area.
}
\examples{
\dontrun{
# load metallo sampling data and add a third field containing a specimen ID
data(metallo)
metallo$id <- factor(paste("Metallo",1:38))

# load NW Atlantic data, convert to class bathy, and plot
data(nw.atlantic)
atl <- as_bathy(nw.atlantic)
plot(atl, deep=-8000, shallow=0, step=1000, col="grey")

# once the map is plotted, use get_sample to get sampling info!
get_sample(atl, metallo, 1, 2)
# click twice
}
}
\seealso{
\code{\link{read_bathy}}, \code{\link{summary_bathy}}, \code{\link{nw.atlantic}}, \code{\link{metallo}}
}
\author{
Eric Pante
}
