% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hawaii}
\alias{hawaii}
\title{Bathymetric data for Hawaii, USA}
\usage{
data(hawaii)
data(hawaii.sites)
}
\value{
\code{hawaii}: a bathymetric object of class \code{bathy} with 539 rows and 659 columns.
\code{hawaii.sites}: data.frame (6 rows, 2 columns)
}
\description{
Bathymetric object of class \code{bathy} created from NOAA GEODAS data and arbitrary locations around the main Hawaiian islands.
}
\details{
\code{hawaii} contains data imported from the NOAA Grid Extract webpage (\url{https://www.ncei.noaa.gov/maps/grid-extract/}) and transformed into an object of class \code{bathy} by \code{read_bathy}.
\code{hawaii.sites} is a 2-columns data.frame containing longitude and latitude of 6 locations spread at sea around Hawaii.
}
\examples{
# load hawaii data
	data(hawaii)
	data(hawaii.sites)

# class "bathy"
	class(hawaii)
	summary(hawaii)

\dontrun{
## use of plot_bathy to produce a bathymetric map
# creation of a color palette
	pal <- colorRampPalette(c("black","darkblue","blue","lightblue"))

# Plotting the bathymetry
	plot(hawaii,image=TRUE,draw=TRUE,bpal=pal(100),asp=1,col="grey40",lwd=.7)

# Adding coastline
	require(mapdata)
	map("worldHires",res=0,fill=TRUE,col=rgb(.8,.95,.8,.7),add=TRUE)

# Adding hawaii.sites location on the map
	points(hawaii.sites,pch=21,col="yellow",bg=color_to_alpha("yellow",.9),cex=1.2)
}
}
\seealso{
\code{\link{plot_bathy}}, \code{\link{summary_bathy}}
}
\author{
see \url{https://www.ncei.noaa.gov/maps/grid-extract/}
}
