% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_bathy.R
\name{is_bathy}
\alias{is_bathy}
\title{Test whether an object is of class bathy}
\usage{
is_bathy(xyz)
}
\arguments{
\item{xyz}{three-column data.frame with longitude (x), latitude (y) and depth (z) (no default)}
}
\value{
The function returns \code{TRUE} or \code{FALSE}
}
\description{
Test whether an object is of class bathy
}
\examples{
xyz <- data.frame(
  lon = rep(c(-5, -4, -3), each = 3),
  lat = rep(c(48, 49, 50), times = 3),
  depth = c(-80, -70, -60, -120, -110, -100, -160, -150, -140)
)

is_bathy(xyz)
is_bathy(as_bathy(xyz))
}
\seealso{
\code{\link{as_bathy}}, \code{\link{summary_bathy}}, \code{\link{read_bathy}}
}
\author{
Eric Pante
}
