% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar_antimeridian_box.R
\name{mar_antimeridian_box}
\alias{mar_antimeridian_box}
\alias{antimeridian.box}
\title{Adds a box to maps including antimeridian}
\usage{
mar_antimeridian_box(object, tick.spacing)

antimeridian.box(object, tick.spacing = 20)
}
\arguments{
\item{object}{matrix of class bathy}

\item{tick.spacing}{spacing between tick marks (in degrees, default=20)}
}
\value{
The function adds a box and tick marks to an existing plot which contains the antimeridian line (180 degrees).
}
\description{
Adds a box on maps including the antimeridian (180)
}
\examples{
data(aleutians)

# default plot:
plot(aleutians,n=1)

# plot with corrected box and labels:
plot(aleutians,n=1,axes=FALSE)
mar_antimeridian_box(aleutians, 10)
}
\seealso{
\code{\link{mar_plot_bathy}}
}
\author{
Eric Pante & Benoit Simon-Bouhet
}
