% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar_as_bathy.R
\name{mar_as_bathy}
\alias{mar_as_bathy}
\alias{as.bathy}
\title{Convert to bathymetric data in an object of class bathy}
\usage{
mar_as_bathy(x)

as.bathy(x)
}
\arguments{
\item{x}{Object of \code{RasterLayer} or \code{SpatialGridDataFrame}, or a three-column data.frame with longitude (x), latitude (y) and depth (z) (no default)}
}
\value{
The output of \code{mar_as_bathy} is a matrix of class \code{bathy}, which dimensions and resolution are identical to the original object. The class \code{bathy} has its own methods for summarizing and ploting the data.
}
\description{
Reads either an object of class \code{RasterLayer}, \code{SpatialGridDataFrame} or a three-column data.frame containing longitude (x), latitude (y) and depth (z) data and converts it to a matrix of class bathy.
}
\details{
\code{x} can contain data downloaded from the NOAA GEODAS Grid Translator webpage (http://www.ngdc.noaa.gov/mgg/gdas/gd_designagrid.html) in the form of an xyz table. The function \code{mar_as_bathy} can also be used to transform objects of class \code{raster} (see package \code{raster}) and \code{SpatialGridDataFrame} (see package \code{sp}).
}
\examples{
# load NW Atlantic data
data(nw.atlantic)

# use mar_as_bathy
atl <- mar_as_bathy(nw.atlantic)

# class "bathy"
class(atl)

# summarize data of class "bathy"
summary(atl)
}
\seealso{
\code{\link{mar_summary_bathy}}, \code{\link{mar_plot_bathy}}, \code{\link{mar_read_bathy}}, \code{\link{mar_as_xyz}}, \code{\link{mar_as_raster}}, \code{\link{mar_as_spatial_grid_data_frame}}.
}
\author{
Benoit Simon-Bouhet
}
