% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar_as_raster.R
\name{mar_as_raster}
\alias{mar_as_raster}
\alias{as.raster}
\title{Convert bathymetric data to a raster layer}
\usage{
mar_as_raster(bathy)

as.raster(bathy)
}
\arguments{
\item{bathy}{an object of class \code{bathy}}
}
\value{
An object of class \code{RasterLayer} with the same characteristics as the \code{bathy} object (same longitudinal and latitudinal ranges, same resolution).
}
\description{
Transforms an object of class \code{bathy} to a raster layer.
}
\details{
\code{mar_as_raster} transforms \code{bathy} objects into objects of class \code{RasterLayer} as defined in the \code{raster} package. All methods from the \code{raster} package are thus available for bathymetric data (e.g. rotations, projections...).
}
\examples{
# load Hawaii bathymetric data
data(hawaii)

# use mar_as_raster
r.hawaii <- mar_as_raster(hawaii)

# class "RasterLayer"
class(r.hawaii)

# Summaries
summary(hawaii)
summary(r.hawaii)

# structure of the RasterLayer object
str(r.hawaii)

\dontrun{
# Plots
#require(raster)
plot(hawaii,image=TRUE,lwd=.2)
plot(r.hawaii)
}
}
\seealso{
\code{\link{mar_as_xyz}}, \code{\link{mar_as_bathy}}, \code{\link{mar_as_spatial_grid_data_frame}}
}
\author{
Benoit Simon-Bouhet
}
