% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar_as_spatial_grid_data_frame.R
\name{mar_as_spatial_grid_data_frame}
\alias{mar_as_spatial_grid_data_frame}
\alias{as.SpatialGridDataFrame}
\title{Convert bathymetric data to a spatial grid}
\usage{
mar_as_spatial_grid_data_frame(bathy)

as.SpatialGridDataFrame(bathy)
}
\arguments{
\item{bathy}{an object of class \code{bathy}}
}
\value{
An object of class \code{SpatialGridDataFrame} with the same characteristics as the \code{bathy} object (same longitudinal and latitudinal ranges, same resolution).
}
\description{
Transforms an object of class \code{bathy} to a \code{SpatialGridDataFrame} object.
}
\details{
\code{mar_as_spatial_grid_data_frame} transforms \code{bathy} objects into objects of class \code{SpatialGridDataFrame} as defined in the \code{sp} package. All methods from the \code{sp} package are thus available for bathymetric data (e.g. rotations, projections...).
}
\examples{
# load Hawaii bathymetric data
data(hawaii)

# use mar_as_spatial_grid_data_frame
sp.hawaii <- mar_as_spatial_grid_data_frame(hawaii)

# Summaries
summary(hawaii)
summary(sp.hawaii)

# structure of the SpatialGridDataFrame object
str(sp.hawaii)

# Plots
plot(hawaii,image=TRUE,lwd=.2)
image(sp.hawaii)
}
\seealso{
\code{\link{mar_as_xyz}}, \code{\link{mar_as_bathy}}, \code{\link{mar_as_raster}}
}
\author{
Benoit Simon-Bouhet
}
