% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar_as_xyz.R
\name{mar_as_xyz}
\alias{mar_as_xyz}
\alias{as.xyz}
\title{Convert to xyz format}
\usage{
mar_as_xyz(bathy)

as.xyz(bathy)
}
\arguments{
\item{bathy}{matrix of class \code{bathy}.}
}
\value{
Three-column data.frame with a format similar to xyz files downloaded from the NOAA Grid Extract webpage (\url{https://www.ncei.noaa.gov/maps/grid-extract/}). The first column contains longitude data, the second contains latitude data and the third contains depth/elevation data.
}
\description{
Converts a matrix of class \code{bathy} into a three-column data.frame containing longitude, latitude and depth data.
}
\details{
The use of \code{mar_as_bathy} and \code{mar_as_xyz} allows to swicth back and forth between the long format (xyz) and the wide format of class \code{bathy} suitable for plotting bathymetric maps, computing least cost distances, etc. \code{mar_as_xyz} is especially usefull for exporting xyz files when data are obtained using \code{mar_subset_sql}, i.e. bathymetric matrices of class \code{bathy}.
}
\examples{
# load celt data
data(celt)
dim(celt)
class(celt)
summary(celt)
plot(celt,deep= -300,shallow= -25,step=25)

# use mar_as_xyz
celt2 <- mar_as_xyz(celt)
dim(celt2)
class(celt2)
summary(celt2)
}
\seealso{
\code{\link{mar_as_bathy}}, \code{\link{mar_summary_bathy}}
}
\author{
Benoit Simon-Bouhet
}
