% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar_check_bathy.R
\name{mar_check_bathy}
\alias{mar_check_bathy}
\alias{check.bathy}
\title{Sort bathymetric data matrix by increasing latitude and longitude}
\usage{
mar_check_bathy(x)

check.bathy(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
The output of \code{mar_check_bathy} is an ordered matrix.
}
\description{
Reads a bathymetric data matrix and orders its rows and columns by increasing latitude and longitude.
}
\details{
\code{mar_check_bathy} allows to sort rows and columns by increasing latitude and longitude, which is necessary for ploting with the function \code{image} (package \code{graphics}). \code{mar_check_bathy} is used within the \code{marmap} functions \code{mar_read_bathy} and \code{mar_as_bathy} (it is also used in \code{mar_get_noaa_bathy} through \code{mar_as_bathy}).
}
\examples{
matrix(1:100, ncol=5, dimnames=list(20:1, c(3,2,4,1,5))) -> a
mar_check_bathy(a)
}
\seealso{
\code{\link{mar_read_bathy}}, \code{\link{mar_as_bathy}}, \code{\link{mar_get_noaa_bathy}}
}
\author{
Eric Pante
}
