% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar_collate_bathy.R
\name{mar_collate_bathy}
\alias{mar_collate_bathy}
\alias{collate.bathy}
\title{Collates two bathy matrices with data from either sides of the antimeridian}
\usage{
mar_collate_bathy(east,west)

collate.bathy(east, west)
}
\arguments{
\item{east}{matrix of class \code{bathy} with eastern data (West of antimeridian)}

\item{west}{matrix of class \code{bathy} with western data (East of antimeridian)}
}
\value{
A single matrix of class \code{bathy} that can be interpreted by \code{mar_plot_bathy}. When plotting collated data (with longitudes 0 to 180 and 180 to 360 degrees), plots can be modified to display the conventional coordinate system (with longitudes 0 to 180 and -180 to 0 degrees) using function \code{mar_antimeridian_box()}.
}
\description{
Collates two bathy matrices, one with longitude 0 to 180 degrees East, and the other with longitude 0 to 180 degrees West
}
\details{
This function is meant to be used with \code{mar_read_bathy()} or \code{mar_read_gebco_bathy()}, when data is downloaded from either sides of the antimeridian line (180 degrees longitude). If, for example, data is downloaded from GEBCO for longitudes of 170E-180 and 180-170W, \code{mar_collate_bathy()} will create a single matrix of class \code{bathy} with a coordinate system going from 170 to 190 degrees longitude.

\code{mar_get_noaa_bathy()} deals with data from both sides of the antimeridian and does not need further processing with \code{mar_collate_bathy()}.
}
\examples{
## faking two datasets using aleutians, for this example
## "a" and "b" simulate two datasets downloaded from GEBCO, for ex.
	data(aleutians)
	aleutians[1:181,] -> a ; "bathy" -> class(a)
	aleutians[182:601,] -> b ; "bathy" -> class(b)
	-(360-as.numeric(rownames(b))) -> rownames(b)

## check these objects with summary(): pay attention of the Longitudinal range
	summary(aleutians)
	summary(a)
	summary(b)

## merge datasets:
	mar_collate_bathy(a,b) -> collated
	summary(collated) # should be identical to summary(aleutians)
}
\seealso{
\code{\link{mar_get_noaa_bathy}}, \code{\link{mar_summary_bathy}}, \code{\link{mar_plot_bathy}}, \code{\link{mar_antimeridian_box}}
}
\author{
Eric Pante
}
