% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar_diag_bathy.R
\name{mar_diag_bathy}
\alias{mar_diag_bathy}
\alias{diag.bathy}
\title{Finds matrix diagonal for non-square matrices}
\usage{
mar_diag_bathy(mat,coord=FALSE)

diag.bathy(mat, coord = FALSE)
}
\arguments{
\item{mat}{a data matrix}

\item{coord}{whether of not to output the coordinates of the diagonal (default is \code{FALSE})}
}
\value{
A vector of diagonal values is \code{coord} is \code{FALSE}, or a table of diagonal coordinates if\code{coord} is \code{FALSE}
}
\description{
Finds either the values of the coordinates of the non-linear diagonal of non-square matrices.
}
\details{
mar_diag_bathy gets the values or coordinates from the first element of a matrix to its last elements. If the matrix is non-square, that is, its number of rows and columns differ, mar_diag_bathy computes an approximate diagonal.
}
\examples{
# a square matrix: mar_diag_bathy behaves as diag
	matrix(1:25, 5, 5) -> a ; a
	diag(a)
	mar_diag_bathy(a)

# a non-square matrix: mar_diag_bathy does not behaves as diag
	matrix(1:15, 3, 5) -> b ; b
	diag(b)
	mar_diag_bathy(b)

# output the diagonal or its coordinates:
	rownames(b) <- seq(32,35, length.out=3)
	colnames(b) <- seq(-100,-95, length.out=5)
	mar_diag_bathy(b, coord=FALSE)
	mar_diag_bathy(b, coord=TRUE)
}
\seealso{
\code{\link{mar_get_transect}}, \code{\link{diag}}
}
\author{
Eric Pante
}
