% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar_is_bathy.R
\name{mar_is_bathy}
\alias{mar_is_bathy}
\alias{is.bathy}
\title{Test whether an object is of class bathy}
\usage{
mar_is_bathy(xyz)

is.bathy(xyz)
}
\arguments{
\item{xyz}{three-column data.frame with longitude (x), latitude (y) and depth (z) (no default)}
}
\value{
The function returns \code{TRUE} or \code{FALSE}
}
\description{
Test whether an object is of class bathy
}
\examples{
# load NW Atlantic data
data(nw.atlantic)

# test class "bathy"
mar_is_bathy(nw.atlantic)

# use mar_as_bathy
atl <- mar_as_bathy(nw.atlantic)

# class "bathy"
class(atl)
mar_is_bathy(atl)

# summarize data of class "bathy"
summary(atl)
}
\seealso{
\code{\link{mar_as_bathy}}, \code{\link{mar_summary_bathy}}, \code{\link{mar_read_bathy}}
}
\author{
Eric Pante
}
