% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar_plot_buffer.R
\name{plot.buffer}
\alias{plot.buffer}
\alias{mar_plot_buffer}
\title{Plots a circular buffer and or its outline}
\usage{
\method{plot}{buffer}(x, outline = TRUE, add = TRUE, ...)

mar_plot_buffer(x, outline = TRUE, add = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{buffer} as produced by the \code{mar_create_buffer()} function.}

\item{outline}{Should the outline of the buffer be plotted (default) or the bathymetric/hypsometric data within the buffer.}

\item{add}{Should the plot be added on top of an existing bathymetric/hypsometric plot (default) or as a new plot}

\item{...}{Further arguments to be passed to the \code{symbols()} function from the \code{graphics} package when \code{outline = TRUE} (default) or to \code{mar_plot_bathy()} when \code{outline = FALSE}.}
}
\value{
Either a plot of the outline of a buffer (default) or a bathymetric map with isobaths of a buffer when \code{outline = FALSE}
}
\description{
\code{mar_plot_buffer} is a generic function that allows the plotting of objects of class \code{buffer}, either as new plots or as a new layer added on top of an existing one. The plotting of both the bathymetry/hypsometry as well as the outline of the buffer is possible.
}
\examples{
# load and plot a bathymetry
data(florida)
plot(florida, lwd = 0.2)
plot(florida, n = 0, lwd = 0.7, add = TRUE)

# add points around which a buffer will be computed
loc <- data.frame(-80, 26)
points(loc, pch = 19, col = "red")

# compute buffer
buf <- mar_create_buffer(florida, loc, radius=1.5)

# plot buffer bathymetry
plot(buf, outline=FALSE, n=10, lwd=.5, col=2)

# add buffer outline
plot(buf, lwd=.7, fg=2)
}
\seealso{
\code{\link{mar_create_buffer}}, \code{\link{mar_combine_buffers}}, \code{\link{mar_plot_bathy}}
}
\author{
Benoit Simon-Bouhet
}
