% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar_plot_profile.R
\name{mar_plot_profile}
\alias{mar_plot_profile}
\alias{plotProfile}
\title{Ploting bathymetric data along a transect or path}
\usage{
mar_plot_profile(profile,shadow=TRUE,xlim,ylim,col.sea,col.bottom,xlab,ylab, \dots)

plotProfile(
  profile,
  shadow = TRUE,
  xlim,
  ylim,
  col.sea,
  col.bottom,
  xlab,
  ylab,
  ...
)
}
\arguments{
\item{profile}{4-columns matrix obtained from \code{mar_get_transect} with argument \code{dist=TRUE}, or from \code{mar_path_profile}.}

\item{shadow}{logical. Should the depth profile cast a shadow over the plot background?}

\item{xlim}{numeric vector of length 2 giving the x coordinate range. If unspecified, values are based on the length of the transect or path.}

\item{ylim}{numeric vector of length 2 giving the y coordinate range. If unspecified, values are based on the depth range of the bathymetric matrix \code{bathy}.}

\item{col.sea}{color for the sea area of the plot. Defaults to \code{rgb(130/255,180/255,212/255)}}

\item{col.bottom}{color for the bottom area of the plot. Defaults to \code{rgb(198/255,184/255,151/255)}}

\item{xlab}{title for the x axis. If unspecified, \code{xlab="Distance from start of transect (km)"}.}

\item{ylab}{title for the y axis. If unspecified, \code{ylab="Depth (m)"}.}

\item{\dots}{arguments to be passed to methods, such as \link{graphical parameters} (see \code{\link{par}})}
}
\value{
a bivariate plot of depth against the kilometric distance from the starting point of a transect or least cost path.
}
\description{
Plots the depth/altitude along a transect or path
}
\examples{
# Example 1:
	data(celt)
	layout(matrix(1:2,nc=1),height=c(2,1))
	par(mar=c(4,4,1,1))
	plot(celt,n=40,draw=TRUE)
	points(c(-6.34,-5.52),c(52.14,50.29),type="o",col=2)

	tr <- mar_get_transect(celt, x1 = -6.34, y1 = 52.14, x2 = -5.52, y2 = 50.29, distance = TRUE)
	mar_plot_profile(tr)

# Example 2:
	layout(matrix(1:2,nc=1),height=c(2,1))
	par(mar=c(4,4,1,1))
	plot(celt,n=40,draw=TRUE)
	points(c(-5,-6.34),c(49.8,52.14),type="o",col=2)

	tr2 <- mar_get_transect(celt, x1 = -5, y1 = 49.8, x2 = -6.34, y2 = 52.14, distance = TRUE)
	mar_plot_profile(tr2)

# Example 3: click several times on the map and press ESC
\dontrun{
	layout(matrix(1:2,nc=1),height=c(2,1))
	par(mar=c(4,4,1,1))
	data(florida)
	plot(florida,image=TRUE,dra=TRUE,land=TRUE,n=40)

	out <- mar_path_profile(as.data.frame(locator(type="o",col=2,pch=19,cex=.8)),florida)
	mar_plot_profile(out)
}
}
\seealso{
\code{\link{mar_path_profile}}, \code{\link{mar_plot_bathy}}
}
\author{
Benoit Simon-Bouhet
}
