% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar_transition_matrix.R
\name{mar_transition_matrix}
\alias{mar_transition_matrix}
\alias{trans.mat}
\title{Transition matrix}
\usage{
mar_transition_matrix(bathy,min.depth=0,max.depth=NULL)

trans.mat(bathy, min.depth = 0, max.depth = NULL)
}
\arguments{
\item{bathy}{A matrix of class \code{bathy}.}

\item{min.depth}{numeric. Minimum depth for possible paths. The default \code{min.depth = 0} indicates that paths can start at sea level.}

\item{max.depth}{numeric or \code{NULL}. Maximum depth for possible paths. The default \code{max.depth = NULL} indicates that paths can extend to the maximum depth of \code{bathy}. See Details.}
}
\value{
A transition object.
}
\description{
Creates a transition object to be used by \code{mar_least_cost_distance} to compute least cost distances between locations.
}
\details{
\code{mar_transition_matrix} creates a transition object usable by \code{mar_least_cost_distance} to computes least cost distances between a set of locations. \code{mar_transition_matrix} rely on the function \code{raster} from package \code{raster} (Hijmans & van Etten, 2012. \url{https://CRAN.R-project.org/package=raster}) and on \code{transition} from package \code{gdistance} (van Etten, 2011. \url{https://CRAN.R-project.org/package=gdistance}).

The transition object contains the probability of transition from one cell of a bathymetric grid to adjacent cells and depends on user defined parameters. \code{mar_transition_matrix} is especially usefull when least cost distances need to be calculated between several locations at sea. The default values for \code{min.depth} and \code{max.depth} ensure that the path computed by \code{mar_least_cost_distance} will be the shortest path possible at sea avoiding land masses. The path can be constrained to a given depth range by setting manually \code{min.depth} and \code{max.depth}. For instance, it is possible to limit the possible paths to the continental shelf by setting \code{max.depth=-200}. Inaccuracies of the bathymetric data can occasionally result in paths crossing land masses. Setting \code{min.depth} to low negative values (e.g. -10 meters) can limit this problem.

\code{mar_transition_matrix} takes also advantage of the function \code{geoCorrection} from package \code{gdistance} (van Etten, 2012. \url{https://CRAN.R-project.org/package=gdistance}) to take into account map distortions over large areas.
}
\examples{
# Load and plot bathymetry
data(hawaii)
summary(hawaii)
plot(hawaii)

\dontrun{
# Compute transition object with no depth constraint
trans1 <- mar_transition_matrix(hawaii)

# Compute transition object with minimum depth constraint:
# path impossible in waters shallower than -200 meters depth
trans2 <- mar_transition_matrix(hawaii,min.depth=-200)

# Visualizing results
par(mfrow=c(1,2))
plot(raster(trans1), main="No depth constraint")
plot(raster(trans2), main="Constraint in shallow waters")
}
}
\references{
Jacob van Etten (2011). gdistance: distances and routes on geographical grids. R package version 1.1-2.  \url{https://CRAN.R-project.org/package=gdistance}
Robert J. Hijmans & Jacob van Etten (2012). raster: Geographic analysis and modeling with raster data. R package version 1.9-92. \url{https://CRAN.R-project.org/package=raster}
}
\seealso{
\code{\link{mar_least_cost_distance}}, \code{\link{hawaii}}
}
\author{
Benoit Simon-Bouhet
}
