% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marmap2-package.R
\docType{package}
\name{marmap2-package}
\alias{marmap2}
\alias{marmap2-package}
\title{marmap2: Import, Plot, and Analyze Bathymetric and Topographic Data}
\description{
marmap2 is an experimental modernization of marmap, a package designed for
downloading, converting, and plotting bathymetric and topographic data in R.
}
\details{
The current development version focuses on a compact modern core:
NOAA/GEBCO import, tibble-based workflows, conversion helpers, \code{sf}/\code{terra}
interoperability, and \code{ggplot2}-based bathymetric maps.
}
\section{Reference}{

Pante E, Simon-Bouhet B (2013). marmap: A Package for Importing, Plotting
and Analyzing Bathymetric and Topographic Data in R. \emph{PLoS ONE}, 8(9),
e73051. \doi{10.1371/journal.pone.0073051}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/besibo/marmap2}
  \item \url{https://besibo.github.io/marmap2/}
  \item Report bugs at \url{https://github.com/besibo/marmap2/issues}
}

}
\author{
\strong{Maintainer}: Benoît Simon-Bouhet \email{besibo@gmail.com}

Authors:
\itemize{
  \item Benoît Simon-Bouhet \email{besibo@gmail.com}
  \item Eric Pante
}

}
\keyword{internal}
