% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nw.atlantic}
\alias{nw.atlantic}
\title{Bathymetric data for the North West Atlantic}
\usage{
data(nw.atlantic)
}
\value{
A three-columns data.frame containing longitude, latitude and depth/elevation data.
}
\description{
Data imported from the NOAA GEODAS server
}
\details{
Data imported from the NOAA Grid Extract webpage (\url{https://www.ncei.noaa.gov/maps/grid-extract/}). To prepare data from NOAA, fill the custom grid form, and choose "XYZ (lon,lat,depth)" as the "Output Grid Format", "No Header" as the "Output Grid Header", and either of the space, tab or comma as the column delimiter (either can be used, but "comma" is the default import format of \code{read_bathy}). Choose "omit empty grid cells" to reduce memory usage.
}
\examples{
# load NW Atlantic data
data(nw.atlantic)

# use as_bathy
atl <- as_bathy(nw.atlantic)

# class "bathy"
class(atl)
summary(atl)

# test plot_bathy
plot(atl, deep=-8000, shallow=-1000, step=1000)
}
\seealso{
\code{\link{plot_bathy}}, \code{\link{summary_bathy}}
}
\author{
see \url{https://www.ncei.noaa.gov/maps/grid-extract/}
}
