% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bathy.R
\name{read_bathy}
\alias{read_bathy}
\title{Read bathymetric data in XYZ format}
\usage{
read_bathy(xyz, header = FALSE, sep = ",", ...)
}
\arguments{
\item{xyz}{three-column table with longitude (x), latitude (y) and depth (z) (no default)}

\item{header}{whether this table has a row of column names (default = FALSE)}

\item{sep}{character separating columns, (default=",")}

\item{...}{further arguments to be passed to \code{read.table()}}
}
\value{
The output of \code{read_bathy} is a matrix of class \code{bathy}. Its
dimensions depend on the resolution and extent of the input xyz table.
}
\description{
Reads a three-column table containing longitude (x), latitude (y) and depth (z) data.
}
\examples{
xyz <- data.frame(
  lon = rep(c(-5, -4, -3), each = 3),
  lat = rep(c(48, 49, 50), times = 3),
  depth = c(-80, -70, -60, -120, -110, -100, -160, -150, -140)
)

tmp <- tempfile(fileext = ".csv")
write.table(xyz, tmp, sep = ",", quote = FALSE, row.names = FALSE)
bathy <- read_bathy(tmp, header = TRUE)
class(bathy)
}
\seealso{
\code{\link{summarise_bathy}}, \code{\link{as_bathy}},
\code{\link{bathy_to_tbl}}
}
\author{
Eric Pante
}
