% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bathy.R
\name{summary_bathy}
\alias{summary_bathy}
\title{Summary of bathymetric data of class \code{bathy}}
\usage{
summary_bathy(object, ...)
}
\arguments{
\item{object}{object of class \code{bathy}}

\item{\dots}{additional arguments affecting the summary produced (see \code{base} function \code{summary}).}
}
\value{
Information on the geographic bounds of the dataset (minimum and maximum latitude and longitude), resolution of the matrix in minutes, statistics on the depth data (e.g. min, max, median...), and a preview of the data.
}
\description{
Summary of bathymetric data of class \code{bathy}. Provides geographic bounds and resolution (in minutes) of the dataset, statistics on depth data, and a preview of the bathymetric matrix.
}
\examples{
xyz <- data.frame(
  lon = rep(c(-5, -4, -3), each = 3),
  lat = rep(c(48, 49, 50), times = 3),
  depth = c(-80, -70, -60, -120, -110, -100, -160, -150, -140)
)

bathy <- as_bathy(xyz)
summary(bathy)
}
\seealso{
\code{\link{read_bathy}}, \code{\link{as_bathy}},
\code{\link{bathy_to_tbl}}
}
\author{
Eric Pante and Benoit Simon-Bouhet
}
