% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_bathy.R
\name{as_bathy}
\alias{as_bathy}
\title{Convert to bathymetric data in an object of class bathy}
\usage{
as_bathy(x)
}
\arguments{
\item{x}{Object of \code{RasterLayer} or \code{SpatialGridDataFrame}, or a three-column data.frame with longitude (x), latitude (y) and depth (z) (no default)}
}
\value{
The output of \code{as_bathy} is a matrix of class \code{bathy}, which dimensions and resolution are identical to the original object. The class \code{bathy} has its own methods for summarizing and ploting the data.
}
\description{
Reads either an object of class \code{RasterLayer}, \code{SpatialGridDataFrame} or a three-column data.frame containing longitude (x), latitude (y) and depth (z) data and converts it to a matrix of class bathy.
}
\details{
\code{x} can contain data downloaded from the NOAA GEODAS Grid Translator webpage (http://www.ngdc.noaa.gov/mgg/gdas/gd_designagrid.html) in the form of an xyz table. The function \code{as_bathy} can also be used to transform objects of class \code{raster} (see package \code{raster}) and \code{SpatialGridDataFrame} (see package \code{sp}).
}
\examples{
# load NW Atlantic data
data(nw.atlantic)

# use as_bathy
atl <- as_bathy(nw.atlantic)

# class "bathy"
class(atl)

# summarize data of class "bathy"
summary(atl)
}
\seealso{
\code{\link{summary_bathy}}, \code{\link{plot_bathy}}, \code{\link{read_bathy}}, \code{\link{as_xyz}}, \code{\link{as_raster}}, \code{\link{as_spatial_grid_data_frame}}.
}
\author{
Benoit Simon-Bouhet
}
