% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_xyz.R
\name{as_xyz}
\alias{as_xyz}
\title{Convert to xyz format}
\usage{
as_xyz(bathy)
}
\arguments{
\item{bathy}{matrix of class \code{bathy}.}
}
\value{
Three-column data.frame with a format similar to xyz files downloaded from the NOAA Grid Extract webpage (\url{https://www.ncei.noaa.gov/maps/grid-extract/}). The first column contains longitude data, the second contains latitude data and the third contains depth/elevation data.
}
\description{
Converts a matrix of class \code{bathy} into a three-column data.frame containing longitude, latitude and depth data.
}
\details{
The use of \code{as_bathy} and \code{as_xyz} allows switching back and forth
between the long xyz format and the historical matrix format of class
\code{bathy}.
}
\examples{
xyz <- data.frame(
  lon = rep(c(-5, -4, -3), each = 3),
  lat = rep(c(48, 49, 50), times = 3),
  depth = c(-80, -70, -60, -120, -110, -100, -160, -150, -140)
)

bathy <- as_bathy(xyz)
as_xyz(bathy)
}
\seealso{
\code{\link{as_bathy}}, \code{\link{bathy_to_tbl}},
\code{\link{tbl_to_bathy}}, \code{\link{summary_bathy}}
}
\author{
Benoit Simon-Bouhet
}
