% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_bathy.R
\name{is_bathy}
\alias{is_bathy}
\title{Test whether an object is of class bathy}
\usage{
is_bathy(xyz)
}
\arguments{
\item{xyz}{three-column data.frame with longitude (x), latitude (y) and depth (z) (no default)}
}
\value{
The function returns \code{TRUE} or \code{FALSE}
}
\description{
Test whether an object is of class bathy
}
\examples{
# load NW Atlantic data
data(nw.atlantic)

# test class "bathy"
is_bathy(nw.atlantic)

# use as_bathy
atl <- as_bathy(nw.atlantic)

# class "bathy"
class(atl)
is_bathy(atl)

# summarize data of class "bathy"
summary(atl)
}
\seealso{
\code{\link{as_bathy}}, \code{\link{summary_bathy}}, \code{\link{read_bathy}}
}
\author{
Eric Pante
}
