% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marmap2-package.R
\docType{package}
\name{marmap2-package}
\alias{marmap2}
\alias{marmap2-package}
\title{Import, plot and analyze bathymetric and topographic data}
\description{
marmap2 is an experimental modernization of marmap, a package designed for downloading, plotting and manipulating bathymetric and topographic data in R. It can query the ETOPO 2022 bathymetry and topography database hosted by the NOAA, use simple latitude-longitude-depth data in ascii format, and take advantage of the advanced plotting tools available in R to build publication-quality bathymetric maps. Functions to query data (bathymetry, sampling information, etc...) are available interactively by clicking on marmap maps. Bathymetric and topographic data can also be used to calculate projected surface areas within specified depth/altitude intervals, and constrain the calculation of realistic shortest path distances.
}
\details{
\tabular{ll}{
Package: \tab marmap2\cr
Type: \tab Package\cr
Version: \tab 0.0.0.9000\cr
}
Import, plot and analyze bathymetric and topographic data
}
\references{
Pante E, Simon-Bouhet B (2013) marmap: A Package for Importing, Plotting and Analyzing Bathymetric and Topographic Data in R. PLoS ONE 8(9): e73051. doi:10.1371/journal.pone.0073051
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/besibo/marmap2}
  \item \url{https://besibo.github.io/marmap2/}
  \item Report bugs at \url{https://github.com/besibo/marmap2/issues}
}

}
\author{
Eric Pante, Benoit Simon-Bouhet and Jean-Olivier Irisson

Maintainer: Benoit Simon-Bouhet \href{mailto:besibo@gmail.com}{besibo@gmail.com}
}
